feat(adminSearchUsers): Move UserSearchRequest into userDetails model file
