--- src/app/core/model/user.model.ts
@@ -1,5 +1,0 @@
-export interface UserSearchRequest {
-  username: string;
-  name: string;
-  emailAddress: string;
-}
